+++ src/main/resources/db/changelog/changelog-master.xml
@@ -3,4 +3,5 @@
     <include file="/db/changelog/changes/changelog-1.0.xml"/>
     <include file="/db/changelog/changes/changelog-1.1.xml"/>
     <include file="/db/changelog/changes/changelog-1.2.xml"/>
+    <include file="/db/changelog/changes/changelog-1.3.xml"/>
 </databaseChangeLog>
