+++ src/main/resources/db/changelog/changes/changelog-1.3.xml
@@ -1,0 +1,24 @@
+<?xml version="1.1" encoding="UTF-8" standalone="no"?>
+<databaseChangeLog xmlns="http://www.liquibase.org/xml/ns/dbchangelog" xmlns:ext="http://www.liquibase.org/xml/ns/dbchangelog-ext" xmlns:pro="http://www.liquibase.org/xml/ns/pro" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.liquibase.org/xml/ns/dbchangelog-ext http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-ext.xsd http://www.liquibase.org/xml/ns/pro http://www.liquibase.org/xml/ns/pro/liquibase-pro-4.1.xsd http://www.liquibase.org/xml/ns/dbchangelog http://www.liquibase.org/xml/ns/dbchangelog/dbchangelog-4.1.xsd">
+    <changeSet id="1" author="Tuan">
+        <insert tableName="class">
+            <column name="id" value="ACCT3603.L12"/>
+            <column name="course_id" value="ACCT3603"/>
+            <column name="teacher_user_id" value="GV0000"/>
+            <column name="start_date" value="2021/01/01"/>
+            <column name="end_date" value="2021/06/01"/>
+            <column name="number_student" value="30"/>
+            <column name="status" value="1"/>
+        </insert>
+
+        <insert tableName="class">
+            <column name="id" value="ADENG1.L12"/>
+            <column name="course_id" value="ADENG1"/>
+            <column name="teacher_user_id" value="GV0000"/>
+            <column name="start_date" value="2021/01/01"/>
+            <column name="end_date" value="2021/06/01"/>
+            <column name="number_student" value="30"/>
+            <column name="status" value="1"/>
+        </insert>
+    </changeSet>
+</databaseChangeLog>
